django/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-colorfield==0.3.2
django-dynamic-preferences==1.10.1
django-nested-admin==3.3.2
django-sortedm2m==3.0.2
django-tinymce==3.0.2
djangorestframework==3.11.1
idna==2.10
persisting-theory==0.2.1
Pillow==7.2.0
python-monkey-business==1.0.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
